Refactored geometry utils
